storage: move chunk impl out of header
