+++ src/v/storage/CMakeLists.txt
@@ -41,6 +41,7 @@ v_cc_library(
     api.cc
     node.cc
     key_offset_map.cc
+    chunk_cache.cc
   DEPS
     Seastar::seastar
     v::bytes
