+++ src/v/storage/chunk_cache.h
@@ -9,18 +9,14 @@
  * by the Apache License, Version 2.0
  */
 #pragma once
+
 #include "base/seastarx.h"
-#include "base/vassert.h"
-#include "resource_mgmt/memory_groups.h"
 #include "ssx/semaphore.h"
 #include "storage/segment_appender_chunk.h"
 
 #include <seastar/core/chunked_fifo.hh>
 #include <seastar/core/future.hh>
-#include <seastar/core/loop.hh>
-#include <seastar/util/later.hh>
-
-#include <boost/iterator/counting_iterator.hpp>
+#include <seastar/core/shared_ptr.hh>
 
 namespace storage::internal {
 
@@ -38,77 +34,25 @@ public:
      */
     static constexpr const alignment alignment{4_KiB};
 
-    chunk_cache() noexcept
-      : _size_target(memory_groups().chunk_cache_min_memory())
-      , _size_limit(memory_groups().chunk_cache_max_memory())
-      , _chunk_size(config::shard_local_cfg().append_chunk_size()) {}
-
+    chunk_cache() noexcept;
     chunk_cache(chunk_cache&&) = delete;
     chunk_cache& operator=(chunk_cache&&) = delete;
     chunk_cache(const chunk_cache&) = delete;
     chunk_cache& operator=(const chunk_cache&) = delete;
     ~chunk_cache() noexcept = default;
 
-    ss::future<> start() {
-        const auto num_chunks = memory_groups().chunk_cache_min_memory()
-                                / _chunk_size;
-        return ss::do_for_each(
-          boost::counting_iterator<size_t>(0),
-          boost::counting_iterator<size_t>(num_chunks),
-          [this](size_t) {
-              auto c = ss::make_lw_shared<chunk>(_chunk_size, alignment);
-              _size_total += _chunk_size;
-              add(c);
-          });
-    }
+    ss::future<> start();
 
-    void add(const chunk_ptr& chunk) {
-        if (_size_available >= _size_target) {
-            _size_total -= _chunk_size;
-            return;
-        }
-        _chunks.push_back(chunk);
-        _size_available += _chunk_size;
-        if (_sem.waiters()) {
-            _sem.signal();
-        }
-    }
+    void add(const chunk_ptr& chunk);
 
-    ss::future<chunk_ptr> get() {
-        // don't steal if there are waiters
-        if (!_sem.waiters()) {
-            return do_get();
-        }
-        return ss::get_units(_sem, 1).then(
-          [this](ssx::semaphore_units) { return do_get(); });
-    }
+    ss::future<chunk_ptr> get();
 
     size_t chunk_size() const { return _chunk_size; }
 
 private:
-    ss::future<chunk_ptr> do_get() {
-        if (auto c = pop_or_allocate(); c) {
-            return ss::make_ready_future<chunk_ptr>(c);
-        }
-        return ss::get_units(_sem, 1).then(
-          [this](ssx::semaphore_units) { return do_get(); });
-    }
+    ss::future<chunk_ptr> do_get();
 
-    chunk_ptr pop_or_allocate() {
-        if (!_chunks.empty()) {
-            auto c = _chunks.front();
-            _chunks.pop_front();
-            _size_available -= _chunk_size;
-            c->reset();
-            return c;
-        }
-        if (_size_total < _size_limit) {
-            auto c = ss::make_lw_shared<chunk>(_chunk_size, alignment);
-            _size_total += _chunk_size;
-            return c;
-        }
-        return nullptr;
-    }
+    chunk_ptr pop_or_allocate();
 
     ss::chunked_fifo<chunk_ptr> _chunks;
     ssx::semaphore _sem{0, "s/chunk-cache"};
@@ -120,9 +64,6 @@ private:
     const size_t _chunk_size{0};
 };
 
-inline chunk_cache& chunks() {
-    static thread_local chunk_cache cache;
-    return cache;
-}
+chunk_cache& chunks();
 
 } // namespace storage::internal
